+++ R/utils_callr.R
@@ -57,6 +57,8 @@ callr_inner <- function(
   targets_arguments,
   targets_options
 ) {
+  getNamespace("targets")$tar_config$set_lock()
+  on.exit(getNamespace("targets")$tar_config$unset_lock())
   withr::local_options(targets_options)
   value <- source(targets_script)$value
   targets_arguments$pipeline <- targets::as_pipeline(value)